# removed pruner function of EquilibriumOptimizer class