Regarding issue #5125 , I removed the ```pruner``` function, and instead we are now only checking if the node exists in the ```fgraph.apply_nodes``` dictionary.