@nouiz 2017-01-19
@ReyhaneAskari did some benchmark with the SBRNN model with a depth of 29. The canonizer was 1.91x faster. Here is some numbers:

Master:
('canonicalize', 'EquilibriumOptimizer', 6, 27724, 17015) - 0.311s
     EquilibriumOptimizer      canonicalize
       time 91.606s for 6 passes

New Commit:
 ('canonicalize', 'EquilibriumOptimizer', 6, 27724, 17015) - 0.301s
     EquilibriumOptimizer      canonicalize
       time 47.869s for 6 passes
--------------------------------------------------------------------------------
Master:
('gpuarray_cut_transfers', 'EquilibriumOptimizer', 3, 35036, 29069) - 0.044s
         EquilibriumOptimizer          gpuarray_cut_transfers
           time 6.153s for 2 passes

New Commit:
('gpuarray_cut_transfers', 'EquilibriumOptimizer', 3, 35036, 29069) - 0.043s
         EquilibriumOptimizer          gpuarray_cut_transfers
           time 1.967s for 2 passes